Refactor dashboard - Split into components (CreateTripDialog, DashboardCard) - Move data that can be moved to separate files
