+++ client/src/components/dashboard/dashboard-card-classes.tsx
@@ -1,0 +1,2 @@
+export const DASHBOARD_CARD_CLASSES =
+  "w-80 text-secondary bg-surface-container-lowest border-3 border-transparent hover:border-primary rounded-lg px-8 pt-5 pb-6 cursor-pointer shadow";
