+++ client/src/mock/mock-archived.ts
@@ -1,0 +1,36 @@
+export const MOCK_ARCHIVED_TRIPS = {
+  2024: [
+    {
+      name: "TRIP NAME #1",
+      destination: "Destination",
+      members: ["Me", "Member"],
+      slug: "1",
+    },
+    {
+      name: "TRIP NAME #2",
+      destination: "Destination",
+      members: ["Me"],
+      slug: "2",
+    },
+    {
+      name: "TRIP NAME #3",
+      destination: "Destination",
+      members: ["Me"],
+      slug: "3",
+    },
+  ],
+  2023: [
+    {
+      name: "TRIP NAME #1",
+      destination: "Destination",
+      members: ["Me"],
+      slug: "4",
+    },
+    {
+      name: "TRIP NAME #2",
+      destination: "Destination",
+      members: ["Me"],
+      slug: "5",
+    },
+  ],
+};
