+++ client/src/types/Trip.ts
@@ -1,0 +1,7 @@
+// Trip as represented in dashboard
+export type Trip = {
+  name: string;
+  destination: string;
+  members: string[];
+  slug: string;
+};
