+++ client/src/views/DashboardView.tsx
@@ -1,67 +1,16 @@
-import { FaPlus } from "react-icons/fa";
 import { useEffect, useState } from "react";
-import {
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-  DialogDescription,
-  DialogTrigger,
-} from "@/components/ui/dialog";
-import { Input } from "@/components/ui/input";
-import { Button } from "@/components/ui/button";
+
 import LogoutButton from "@/components/auth/LogoutButton";
-import { useNavigate } from "react-router-dom";
-import AddCollaborators from "@/components/add-collaborators/AddCollaborators";
 
-type Trip = {
-  name: string;
-  destination: string;
-  members: string[];
-  slug: string;
-};
-
-const archivedTrips = {
-  2024: [
-    {
-      name: "TRIP NAME #1",
-      destination: "Destination",
-      members: "Me, Name1, Name2, Name3, and Name4",
-    },
-    {
-      name: "TRIP NAME #2",
-      destination: "Destination",
-      members: "Me, Name1, Name2, Name3",
-    },
-    {
-      name: "TRIP NAME #3",
-      destination: "Destination",
-      members: "Me, Name1",
-    },
-  ],
-  2023: [
-    {
-      name: "TRIP NAME #1",
-      destination: "Destination",
-      members: "Me, Name1, Name2, Name3, and Name4",
-    },
-    {
-      name: "TRIP NAME #2",
-      destination: "Destination",
-      members: "Me, Name1, Name2, Name3",
-    },
-  ],
-};
+import { Trip } from "@/types/Trip";
+import { MOCK_ARCHIVED_TRIPS } from "@/mock/mock-archived";
+import CreateTripDialog from "@/components/dashboard/CreateTripDialog";
+import DashboardCard from "@/components/dashboard/DashboardCard";
 
 const DashboardView = () => {
-  const navigate = useNavigate();
   const [trips, setTrips] = useState<Trip[]>([]);
-  const [form, setForm] = useState({
-    name: "",
-    destination: "",
-    collaborators: [],
-  });
-  const [open, setOpen] = useState(false);
+  const HEADER_CLASSES =
+    "text-secondary text-xl font-semibold text-center mb-10";
 
   useEffect(() => {
     const fetchTrips = async () => {
@@ -86,7 +35,6 @@ const DashboardView = () => {
         }));
 
         setTrips(formattedTrips);
-        console.log(data);
       } catch (err) {
         console.error("Failed to load trips:", err);
       }
@@ -95,137 +43,47 @@ const DashboardView = () => {
     fetchTrips();
   }, []);
 
-  const setCollaborators = (collaborators: string[]) => {
-    setForm((prev) => ({ ...prev, collaborators }));
-  };
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setForm({ ...form, [e.target.name]: e.target.value });
-  };
-
-  const handleCreate = async () => {
-    const token = localStorage.getItem("token");
-
-    try {
-      const serverUrl = import.meta.env.VITE_SERVER_URL;
-      const res = await fetch(`${serverUrl}/api/trips`, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          Authorization: `Bearer ${token}`,
-        },
-        body: JSON.stringify({
-          tripName: form.name,
-          destinations: form.destination,
-          collaboratorEmails: form.collaborators,
-        }),
-      });
-
-      if (!res.ok) throw new Error("Failed to create trip");
-
-      const newTrip = await res.json();
-
-      const trip: Trip = {
-        name: newTrip.tripName,
-        destination: newTrip.destinations,
-        members: newTrip.users.map(
-          (u: any) => u.name || u.email || "Collaborator"
-        ),
-        slug: newTrip.tripSlug,
-      };
-
-      setTrips((prev) => [...prev, trip]);
-      setForm({ name: "", destination: "", collaborators: [] });
-      setOpen(false);
-    } catch (err) {
-      console.error("Error creating trip:", err);
-    }
-  };
   return (
-    <div className="bg-gray-100 min-h-screen">
-      <nav className="bg-green-700 text-white p-4 flex justify-between items-center">
+    <div className="flex flex-col w-full h-full justify-center bg-surface">
+      <nav className="bg-primary text-on-primary p-4 flex justify-between items-center">
         <h1 className="text-xl font-bold">trip planner</h1>
-        {/* <button className="bg-white text-green-700 px-4 py-2 rounded-md">Sign Out</button> */}
-        <LogoutButton />
+        <LogoutButton className="h-10 w-24 text-on-surface bg-surface hover:bg-surface/90" />
       </nav>
-
-      <div className="w-4xl mx-auto p-6">
-        <h2 className="text-lg font-semibold text-center mb-4">my trips</h2>
-        <div className="grid grid-cols-2 gap-4 justify-center">
-          {/* Create Trip Dialog */}
-          <Dialog open={open} onOpenChange={setOpen}>
-            <DialogTrigger asChild>
-              <div className="bg-white shadow-md rounded-lg flex items-center justify-center h-32 border border-gray-300 cursor-pointer">
-                <FaPlus className="text-3xl text-gray-500" />
-              </div>
-            </DialogTrigger>
-            <DialogContent className="flex flex-col sm:w-md">
-              <DialogHeader>
-                <DialogTitle>Create a new trip</DialogTitle>
-                <DialogDescription>
-                  Start off by adding a few details. You can always change these
-                  later.
-                </DialogDescription>
-              </DialogHeader>
-              <div className="space-y-4">
-                <div>
-                  <label className="block text-sm font-medium text-gray-700">
-                    Trip Name
-                  </label>
-                  <Input
-                    name="name"
-                    value={form.name}
-                    onChange={handleChange}
+      <div className="w-fit flex flex-col items-center self-center gap-20 my-10">
+        <div>
+          <h2 className={HEADER_CLASSES}>my trips</h2>
+          <div className="grid grid-cols-1 sm:grid-cols-2 gap-4 justify-center">
+            <CreateTripDialog setTrips={setTrips} />
+            {trips.map((trip) => (
+              <DashboardCard
+                key={trip.slug}
+                name={trip.name}
+                destination={trip.destination}
+                members={trip.members}
+                slug={trip.slug}
+              />
+            ))}
+          </div>
+        </div>
+        <div>
+          <h2 className={HEADER_CLASSES}>archive</h2>
+          {Object.entries(MOCK_ARCHIVED_TRIPS).map(([year, trips]) => (
+            <div key={year} className="flex flex-col lg:flex-row items-center gap-6 mb-4">
+              <h3 className="text-outline font-semibold">- {year} -</h3>
+              <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4">
+                {trips.map((trip) => (
+                  <DashboardCard
+                    key={trip.slug}
+                    name={trip.name}
+                    destination={trip.destination}
+                    members={trip.members}
+                    slug={trip.slug}
                   />
-                </div>
-                <div>
-                  <label className="block text-sm font-medium text-gray-700">
-                    Destination(s)
-                  </label>
-                  <Input
-                    name="destination"
-                    value={form.destination}
-                    onChange={handleChange}
-                  />
-                </div>
-                <AddCollaborators setCollaborators={setCollaborators} />
-                <div className="flex justify-end pt-2">
-                  <Button onClick={handleCreate}>Create</Button>
-                </div>
+                ))}
               </div>
-            </DialogContent>
-          </Dialog>
-
-          {/* Existing Trips */}
-          {trips.map((trip, index) => (
-            <div
-              key={index}
-              className="bg-white shadow-md rounded-lg p-4 cursor-pointer hover:shadow-lg"
-              onClick={() => navigate(`/trip/${trip.slug}`)}
-            >
-              <h3 className="text-md font-bold">{trip.name}</h3>
-              <p className="text-gray-500">{trip.destination}</p>
-              <p className="text-gray-600 text-sm">{trip.members.join(", ")}</p>
             </div>
           ))}
         </div>
-
-        {/* Archive Section */}
-        <h2 className="text-lg font-semibold text-center mt-8 mb-4">archive</h2>
-        {Object.entries(archivedTrips).map(([year, trips]) => (
-          <div key={year}>
-            <h3 className="text-gray-500 font-semibold mb-2">– {year} –</h3>
-            <div className="grid grid-cols-3 gap-4">
-              {trips.map((trip, index) => (
-                <div key={index} className="bg-white shadow-md rounded-lg p-4">
-                  <h3 className="text-md font-bold">{trip.name}</h3>
-                  <p className="text-gray-500">{trip.destination}</p>
-                  <p className="text-gray-600 text-sm">{trip.members}</p>
-                </div>
-              ))}
-            </div>
-          </div>
-        ))}
       </div>
     </div>
   );
